backend/requirements.txt
flask
flask_cors
pymongo
python-dotenv
opencv-python-headless
numpy
keras
tensorflow
pynput
pygetwindow
python-dateutil
gunicorn